Gradle support for the Java builder
Tool detection preferences are now the following

1. Maven (check for pom.xml)
2. Gradle (check for build.gradle)
3. Ant (no checks, a last resort fallback)

While in theory there are Gradle-based projects that have pom.xml in the repository
(for example, Elastic Search), it is not clear when pom.xml is there just for artifact
publishing or is actually used.

As one more example, Neo4J Server Community Edition seems to have both Maven and
Gradle in repository modules. So we have to pick something and that something
is Maven. It is never a problem to override `script:` and `install:` in `.travis.yml`
to make Travis run what you really want.