Change after_script behavior
In commit 1fd5778 I introduced after_test option for .travis.yml, which
allowed to run command regardless the test result. After reviewing this,
we came up with better solution. Previously, after_script was run only
if script passed, which made it virtually the same as after_success.

This commit changes after_script's behavior to run regardless test
result. Please not that this changes a few things:

* order of running stages is different, after_script runs after success
  failure commands
* fail in after_script will not make test fail - it's run only when test
  result is already known
* it's possible to check test result with TRAVIS_TEST_RESULT env var in
  after_script now