Use :env and :global_env as env vars sources
The ability to add 'global' env vars was added to .travis.yml. This
allows to set environment variables, which will be used in each of the
jobs in the build. There is a problem with handling such construct -
till now travis-build was able to handle only one env source:
config[:env]. Such behavior requires merging global and matrix env
arrays and put them together in job's config. This is the source of
limitations and problems. To solve the problem, this commit adds support
for :global_env key in config array, to allow send matrix env vars and
global env vars as separate objects.