Install Android SDK components on android builder setup
The current commit wraps the common command pattern to install one sdk
component. It assumes that we have the `android` command installed and
on PATH.

Also, it uses the `sdk_components` configuration key to fetch the list
of components to install.

A set of default components is given, and they are the most recent ones
at the date of this commit.