fix(objective-c): start up the simulator at the beginning of the build
Starting the simulator at the beginning of the build causes interaction
with the simulator to be much more stable. While testing this it looked
like the command finished very quickly, since it only tells the Finder
to open the simulator, it does not wait for it to open. Therefore I do
not think it's an issue to run this command for every build.

Close travis-ci/travis-ci#1630.