fix(hosts): update sed expression to work with BSD sed
BSD sed is used on OS X. This version of sed requires an argument to be
passed together with the -i flag. Bash interprets -i'' and -i the exact
same way (this is how string concatenation works in Bash), so the space
is required for sed to notice the empty flag.