Fix shell redirection
Commands such as './some-command 2>&1 > some.log' will end up only
redirecting stdout to the log file, stderr will be redirected to the
"old" stdout. This can be fixed by either changing the order of the
redirections or changing to '&> some.log', which redirects both stdout
and stderr to the log file. I went with '&>' here and changed the other
references as well to make it consistent.