Run `export`s after enabling paranoid mode
One could set up an environment that executes code using $() or ``,
which might use sudo, so the exports should go after paranoid mode is
enabled (if it isn't enabled it doesn't matter what order it goes in).

This also extracts most of the "builtin" stages into a separate
configure stage instead.