Move directory cache setup to `setup` stage
The directory cache inserts some commands un-escaped from the config (mainly
when adding directories that should be cached). Instead of going through the
code and escaping them (I think they are unescaped so environment variables can
be used), I moved the directory cache setup to `setup`, which is after paranoid
mode is enabled, so these commands should be safe to run.