Change timing to use the Bash time keyword
By setting TIMEFORMAT we can get a similar output to what the
travis_time functions gave us without having to deal with TRAVIS_CMD
changing and similar things.

The main difference in output is that we don't get the "start" and "end"
time, but I'd argue that those are not as necessary. A newline will also
always be added to the end of TIMEFORMAT before displaying it, so having
a carriage return is not necessary, and we'll have to deal with this in
travis-web to 'remove' it nicely. It will be displayed in raw logs, but
I think that might actually be useful.