Run `cargo build` in `script` phase
The analog to C builds is “make && make test”.

cargo build and cargo test can in theory have different flags when
compiling (one is compiling the actual library/binary, the other a
special test binary), so both of them should be run. The fact that
things do/don’t compile is a sort-of test in itself, so running them in
`script` seems more appropriate.