Using the production header.sh for `travis run`
so that output is more real, there aren't explosions around usage of
`travis_cmd`, etc.

Without this, I was consistently getting errors around `travis_cmd` and
`travis_fold` not being defined.  I'm totally willing to accept this was
a case of me doin' it wrong, but figured I'd submit this for discussion
at the very least :smiley_cat:.