Allow `perlbrew use` to fail and continue
Apparently, perlbrew changed the behavior in such a way that
when `perlbrew use` fails, it exits with nonzero.
This is correct behavior, but now a lot of users are complaining
about broken builds.

We let it fall through, so that the build can continue.

See also: https://github.com/travis-ci/travis-build/pull/378