Add customizable 'before_cache' stage
Often users want to perform cleanup tasks before the new cache
is uploaded; for example, deleting a new marker file which
fools casher into thinking that the entire cache is new.

Also, `after_result` stage is renamed to `cache`, which correctly
describes the purpose of this stage.
Note that this is now a non-critical stage, since failure to update
cache should not fail the entire job.