Escape all values from users
Before this commit we were not escaping env vars values from users. A
big drawback of that is that our secure env vars are not "secure by
default", becasue if an unaware user sets a password with special
characters, like foo&\, bash will throw an error possibly outputting the
password.

This commit escapes all of the env values with `shellescape`.