Restart MySQL on boot
With the move to MySQL 5.6 packages from mysql.com, we noticed that
MySQL socket (`/var/run/mysqld/mysqld.sock`) is inexplicably lost.
The process is running, and users can connect with TCP
(e.g., `mysql -h 127.0.0.1`), but connexions over the socket are
unsuccessful, leading to lots of frustration.

Upon restarting, though, the socket is re-created, and things
work.

I have not been able to determine what is removing the socket.