Leave CCACHE_DISABLE untouched
Setting CCACHE_DISABLE to `true` in all cases is a bad idea, but
it is turned off only if `cache: ccache` shortcut is used.
This causes a problem with those builds that uses ccache but
does not use `cache: ccache`.

It is best to leave CCACHE_DISABLE untouched.

Fixes https://github.com/travis-ci/travis-ci/issues/3727