Tweak hosts addon
Instead of slurping, editing and dumping the content
in a single pipeline, we construct the file content in one
command, then overwrite the file in another.

The hope here is that this reduces the chance of race conditions
that I suspect is contributing to the empty /etc/hosts we see manifest
as https://github.com/travis-ci/travis-ci/issues/3922.