Use `prove` directly instead of `make` in `script` step
We can't assume that the users will have created a Makefile with appropriate
commands for running the test suite, however, we *can* hope that they have a
`t/` directory with tests.  Also `prove` should be part of the standard
Perl5 installation, so it's ok to rely on this command being available.