Update sbt to latest version
Fixes https://github.com/travis-ci/travis-ci/issues/4527

This, in effect, negates what we do in the cookbooks
https://github.com/eed3si9n/travis-cookbooks/blob/8ef970bff256656219f1673472445c70d44789af/ci_environment/sbt-extras/recipes/default.rb#L16-L22
and makes the build (ever so slightly) more susceptible to
network (but the risk should be minimal, since we are relying on
GitHub anyway).

When the new image containing this version of sbt is provisioned
and deployed, this commit should be either updated or reverted.