Accept 'latest', 'latest-beta' and 'latest-esr' as Firefox versions
Resolves https://github.com/travis-ci/travis-ci/issues/4644

Make a curl call and scrape the HTTP directory listing
to find the latest version of firefox corresponding to:
latest, latest-esr, latest-beta.

It seems that 'latest' and friends are just symbolic
links to the current respective versions.

There *may* be an API endpoint to discover them without
scraping HTML, but I have not found it.