Enforce OTP release requirement for Elixir
Elixir 1.2.0 will require OTP 18.0 or later, while
1.1.x will not run on OTP 18.0.

While it is not impossible to test with both Elixir 1.1.x and
1.2.x in the same build with such requirements, it is a bit
cumbersome, as it will require a hack with `matrix.include`.

This commit allows implicit OTP Release requirement specification
for just such a case.

If OTP release requirement is not met for one reason or another,
we will automatically select (and install if unavailable) it.

This does not allow testing Elixir versions with different OTP
releases, but such cases should be uncommon (and are already using
the build matrix one way or another, so the `matrix.include` trick
should be familiar).