Check for Go 1.5 vendoring before Godeps init.
If we are using Go 1.5's vendoring method, we could have a
Godeps/Godeps.json file and _not_ want to run `godep restore`. Godep is
used as a tool to manage dependencies under Go 1.5's vendoring method,
and `godep restore` is _not_ supposed to be run in those cases.

This fixes that by adding a check to see if we're using Go 1.5's
vendoring method before we check if godep is being used, and skipping
the godep section if we're using 1.5's vendoring method instead.