Alter `default_branches` based only on keys in `app`
Previously, `default_branches` was set by finding the keys of all
Hashes given as a value in the config.

For example, given

```yaml
hash1:
  key1: value1
  key2: value2
hash2:
  key3: value3
  key4: value4
```

we collect `key1`, `key2`, `key3`, `key4`.

This was done in order to deal with the `heroku` provider's `app`
key, which is documented in https://docs.travis-ci.com/user/deployment/heroku#Deploying-Custom-Application-Names

This caused a problem when we want to test `dpl` by installing it from
source (https://github.com/travis-ci/dpl/blob/master/TESTING.md),
as the structure of the `edge` hash caused the `default_branches`
to be `['source', 'branch']`, and the `master` branch need to be
explicitly enabled instead.

We resolve this issue by just looking at `app` key-value pair, since
it is the only one we should be concerned about when setting
`default_branches`.