Special case tip.
When using the tip version of Go, we want to always consider it as after
1.5 (and, more importantly, 1.2). Previously (before
travis-ci/travis-build#658) this was fine, because 'tip' > '1.2'
evaluates true in Ruby. After the patch, we were using true comparisons
instead of string comparisons, and suddenly it was defaulting to
'0.0.1', which means 'tip' > '1.2' effectively starting evaluating
false. Which is not what we want. So now all our comparisons should be
special-casing tip.