Fix case where `branch` value is a Hash
Revert the logic to construct the bash conditional, and instead
tweak the logic to construct the `branches` array.

Since `deploy.branch` is a special case
`(Array(on[:branch] || default_branches)`
for the branch condition, we need to deal with the case where
this has a Hash in it; e.g.,

deploy:
  provider: X
  branch:
    master: foo
    develop: bar
  ⋮

This doesn't seem like a common configuration, but it is allowed
elsewhere (`app`, `environment`).