Allow multiple API auth tokens to be specified in the config
This makes it easier to rotate the tokens when necessary, since you can
add the new token to the config, while keeping the old one valid until
you're done swapping out the token.

This should be backwards compatible, since String#split on a string
without the given separator will just return the string. For example:

    >> "foo".split(",")
    => ["foo"]