[csharp] Fix Mono 4.4.1 commands not being available in PATH, causing failed builds
Mono 4.4.1 switched from installing binaries to /usr/local/bin to adding an entry via /etc/paths.d.
This has the unfortunate consequence that `mono`, `xbuild` etc. won't be available right away in the same shell session
as before (which was the case because /usr/local/bin is in the PATH).

On Travis this then makes the build fail as none of the commands work.
Fixing this by running the path_helper script to update the PATH.