Introduce CleanUpPath appliance
This appliance removes `::` from `$PATH`, so that the current
directory is not on `$PATH`.

See https://github.com/travis-ci/travis-ci/issues/5363

The real fix for the issue is to fix the cookbooks, but this
appliance will do while we wait for the update.
When the update is rolled out, this addon will be a noop, and can
be safely removed.