Activate hhvm-nightly correctly
Resolves https://github.com/travis-ci/travis-ci/issues/6616

`hhvm-nightly` (on Trusty, for sure, but probably on Precise) did not
actually work because there was no `hhvm-nightly` directory under
`$HOME/.phpenv/versions`, where `phpenv` looked for PHP installations.

Furthermore, `phpenv global hhvm` is incorrect for `hhvm-nightly`.
This incorrect invocation led to installation of the `hhvm` package
instead, which reverted the HHVM version to whatever stable release
that package refers to.