Ensure SHELL, TERM, and USER are defined
and run header check in an isolated env, making stronger assertions about env
vars and values.

Background:

A recent rollout of refreshed infrastructure for running container-based builds
revealed the assumption that `SHELL`, `TERM`, and `USER` env vars were being
defined by ... something ... somewhere... and that they would always be around.
As it turns out, when executing a job via the docker remoting API as opposed to
SSH, these vars are not defined.  This addition ensures that they are defined,
with values that *should* work for the use cases of which we're currently aware.