Fix an error where Sierra builds can't use git
Due to the CVE earlier this year regarding "Roaming" mode in SSH, Apple
has decided it's a good idea to just completely take out SSH roaming
entirely, starting in macOS 10.12. Defining an invalid directive in the
SSH configuration breaks Git, so we need to not set that directive.

The "if OS minor version is < 12" syntax is weird, but is used elsewhere
in the repo just the same way.