Show location of system info file(s) instead of dumping
The reasoning behind this is that it is relatively unknown how useful the
contents of the system info file are to our users, and yet every single job
contains a copy of the file.  We pay for that data at multiple layers, including
long term storage in S3.  This change is meant to be somewhat provocative, with
the hope being that users who depend on having the dumped data present make
themselves known.