Add gimme as static file
Update gimme from our static file stash if necessary.
The logic here remains the same; if gimme is not found on the image,
or reinstall is forced, we fetch the file.
If that fails for some reason, we fall back to the previous logic
and fetch it from raw.githubusercontent.com