Always update nvm
If we do not know what version of `nvm` which `travis-build` web app
is serving at the build script compile time (with some constant known
to this file), we must fetch `nvm.sh` from the web app at run time to
find it.
If we *are* fetching `nvm.sh` for any reason, though, we might as
well use it to update nvm.sh.