Look for phpunit in order defined by composer
When composer is used, preferred phpunit may not be the one that
is provided by the PHP runtime we offer.

As described in https://getcomposer.org/doc/articles/vendor-binaries.md#can-vendor-binaries-be-installed-somewhere-other-than-vendor-bin-
(and as discovered by an additional experiment), the order is:
1. $COMPOSER_BIN_DIR/phpunit
2. `config: "bin-dir"` in composer.json
3. vendor/bin/phpunit

If none of these applies, we invoke `phpunit`, which is the one
provided by our PHP runtime.