Turn build stages into triples
The third field now indicates whether or not the stage should actually
run when we are running the debug build.

*All* stages are defined, but only some will be *executed*.

:always => it executes whether or not this is a debug build
true => executes if and only if this is a debug build
false => executes if and only if this is *not* a debug build

One spec no longer applies, and is therefore removed.