Use `#run_stage` on deployment section
This avoids nesting here-docs in the deployment portion in the resulting
bash script.

Previously, these deployment portion was called from the `after_success`
section, which meant that the here-doc end markers in the deployment
portions had leading white spaces, invalidating the otherwise well-formed
marker.

This manifested in errors such as:

```
/home/travis/.travis/job_stages: line 670: warning: here-document at line 628 delimited by end-of-file (wanted `EOFUNC')
/home/travis/.travis/job_stages: line 671: syntax error: unexpected end of file
```

The affected function, and the functions that follow are therefore undefined:

```
/home/travis/build.sh: line 754: travis_run_after_success: command not found
/home/travis/build.sh: line 755: travis_run_after_failure: command not found
/home/travis/build.sh: line 756: travis_run_after_script: command not found
/home/travis/build.sh: line 757: travis_run_finish: command not found
```