Drop only `/` and encode branch name
By allowing a wider set of characters in the branch name to pass
to casher, we allow emojis in the branch name: "🐡"

Technically speaking, it is not necessary to drop `/`, since it is
a legal character in git branch names.
However, retaining this common character in the cache name will
invalidate existing caches.