Set product in each `case`
Since we are making an extensive exception in the
`latest-unsigned` case, it reads better if we set the `product`
value here.
Otherwise, `product` would be set to the `source_url` value;
while this assignment is harmless (because it is not used later),
not setting is superior because of clarity of intentions.