Set up RVM aliases when one is required
The build image may not know about the MAJOR.MINOR Ruby version
aliases, so we ensure that they exist.

Ideally, RVM_VERSION_ALIASES should be discovered at the build time
(or `travis-build` deploy time), but it is a surprisingly difficult
problem to solve reliably at the moment. So we simply hard code the
known releases for now.