Move browserstack install/start logic to methods
The use of `return` in this context is used as "skip the rest of
the logic for browserstack fold", but by using Ruby's `return`,
it actually breaks the resulting Bash script (with a syntax error).

By moving this portion into a method call, the use of "return" is
correctly reflected in the resulting valid Bash script.