Run `rvm use` in configure
RVM resets certain environment variables to ensure that it can
set up Ruby environment correctly.
One such environment variable is $BUNDLE_PATH. If `.travis.yml`
attempts to set $BUNDLE_PATH in env, this is no longer effective
in the subsequent portions of the build. This is confusing.

We resolve this by moving `rvm …` commands to configure, so that
`export` is effective.