Keep `rvm current` STDOUT
When setting up `casher`, ensure that we retain the output from `rvm
current`. If we sent it to `/dev/null`, we effectively run `rvm use`,
which will respect `.ruby-version` if it is present, and if this
version is pre-installed, that version is selected, even though we had
previously switched to a different version.

This is the cause of https://github.com/travis-ci/travis-ci/issues/8069