Skip update_static_files Rake task during the required jobs
This task is sensitive to GitHub API rate limit, and often fails
during the external PR builds because of it.

We move it to a separate job, which is allowed to fail. This has
an added benefit of not running the Docker build stage during the
external PR builds, because the new "test rake" stage will fail
then.