Fix Go version check for go get -t
gimme recently added the ability to request 'stable' instead of tip or
master. However, if one requests stable, go get -t is disabled despite
the fact that it would be enabled on the actual version (1.8.3 at time
of writing). Since comparable_go_version returns 0.0.1 if the version
isn't a number, we get the options for < 1.2.

Instead we should treat all non-numbered go versions the same, which
means using go get -t.