Enforce tag presence on GitHub Releases deployment
In the documentation, we state that the GitHub Releases deployment
requires a tag build, but this is not true.

We happily oblige, and in the process, creates an unwanted release,
based on an unexpected tag name (untagged-HEX); this results in
https://github.com/travis-ci/travis-ci/issues/8185.

This commit resolves that issue by triggering the tags condition
when the deployment provider is `release`.