Move Rust install step to after cache is set up
Rust sets up its binaries in `$HOME/.cargo`, so if cache is in use,
certain binaries cached there will remain that of old installation,
which could be a surprise.

By moving the installation logic to after the cache, we ensure that
rustup-init.sh result is not overwritten by the cache.