Install different versions of dpl based on travis_internal_ruby output
This is the first step in moving `dpl` to a better place, where
we can drop MRI 1.9 support.

We plan to release `dpl` 1.9 in the not-so-distant future, which will
require Ruby >= 2.2.
If `travis_internal_ruby` returns 1.9, then, we need to install
some 1.8.x.

This PR also adds a new helper method for specs, `sexp_if_then_else`,
which looks for the `if` s-expression for the given condition, then
returns the hash naturally mapping to it, with keys `:then` and
`:else`.
This allows a cleaner and more robust search for the s-expression
we want for the specs.