Rescue "Malformed version number" when R## is specified
OTP Release versions such as `R16B03-1` is not supported
by Gem::Version, so we rescue them, and assert that the version
requirement is not met.

This may not be strictly speaking accurate; it is possible that
some older Elixir versions *could* still run on OTP Releases
older than 17.0 (with the last one before it being R16B03-1).
The 17.0 release is already a few years old, and we will assume that
it is not desirable to run with these R* releases.