Let `cache.name` override cache slug name we compute
So that the jobs within a build can share caches arbitrarily.

It is not possible to put this change in
`Travis::Build::Script#cache_slug`, since multiple modules may call
this with `super`; in such cases the end cache name will be different
from what is given in `cache.name`.