julia: Only run `Pkg.clone(pwd())` on < `0.7-`
Recent changes to Julia 0.7 have obviated the need for a `Pkg.clone()` step; so if there is a `Project.toml` file in the current directory, and we're running on Julia 0.7+, then don't bother to `Pkg.clone(pwd())`.  If there is no `Project.toml` file, or if we're running on Julia 0.6, then we should still `Pkg.clone(pwd())`.