Return `data` instead of `Enumerable#compact!` result
`Enumerable#compact!` returns `nil` if there is no `nil` to remove.
This may result in removing everything from `env_vars`.
Instead, return the result from `Enumerable#compact`.

I _think_ it is OK to not modify `data.env_vars` here (which
`Enumerable#compact!` did), as this is the only place it is accessed.