Emit empty bash statement from Noop dir cache
When directory cache is not configured, we could end up calling various
Noop directory_cache methods inside bash `if` and `else` by itself.
This could lead to invalid bash constructs such as:

    travis_fold start cache.npm
    echo
    if [[ *condition* ]]; then
    else
    fi
    travis_fold end cache.npm

    :
    }

By implementing certain directory_cache methods to emit empty bash
statement, we avoid this error.

It is also possible to *always* consider this possibility when using
directory_cache, but that would be too much burden on the future
developers.