fix elm versioning for 0.19.1 plus
The versioning scheme has changed with the release of elm 0.19.0.
Instead of npm tags of the form: elmMAJOR.MINOR.PATCH elm npm
packages use tags of the form: latest-MAJOR.MINOR.PATCH.

This commit updates travis-builds logic to pick the correct prefix.
It will default to using the "latest-" prefix unless the config
it towards the "elm" prefix.